Initial commit of Java files.
