+++ java/src/main/java/phevaluator/BatchEvaluator.java
@@ -1,4 +1,4 @@
-package evaluator;
+package phevaluator;
 
 import java.util.ArrayList;
 import java.util.List;
