+++ java/src/main/java/phevaluator/Card.java
@@ -1,4 +1,4 @@
-package evaluator;
+package phevaluator;
 
 import java.util.Arrays;
 import java.util.stream.Collectors;
