+++ java/src/main/java/phevaluator/DPTables.java
@@ -1,4 +1,4 @@
-package evaluator;
+package phevaluator;
 
 class DPTables{
