+++ java/src/main/java/phevaluator/Evaluator.java
@@ -1,4 +1,4 @@
-package evaluator;
+package phevaluator;
 
 public class Evaluator{
