+++ java/src/main/java/phevaluator/Hash.java
@@ -1,4 +1,4 @@
-package evaluator;
+package phevaluator;
 
 class Hash{
