+++ java/src/main/java/phevaluator/HashTable.java
@@ -1,4 +1,4 @@
-package evaluator;
+package phevaluator;
 
 import java.io.IOException;
 import java.nio.file.Files;
