+++ java/src/test/java/phevaluator/IdTestInputBean.java
@@ -1,4 +1,4 @@
-package evaluator;
+package phevaluator;
 
 import com.opencsv.bean.CsvBindByName;
